Updated README.md
Modder here. I've been making a mod similar to this, and I noticed a few inaccuracies. For example, discussion time is a 16 bit value, not 8 (which is surprising, as only 8 bit values are normally used. This means that a truly insane person could make a discussion time of 65535 seconds, or a little more than 18 hours. I also updated it to mention the Airship map.